% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pshd.R
\name{estimate_pshd}
\alias{estimate_pshd}
\title{Find a netwwork structure that minimizes the expected PSHD loss among a set of samples}
\usage{
estimate_pshd(samples, a = 1)
}
\arguments{
\item{a}{}
}
\value{
xxxxx
}
\description{
Find a netwwork structure that minimizes the expected PSHD loss among a set of samples
}
