% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pshd.R
\name{pshd}
\alias{pshd}
\title{Penalized Structural Hamming Distance}
\usage{
pshd(g1, g2, a = 1)
}
\arguments{
\item{g1}{A \eqn{p \times p} matrix}

\item{a}{}
}
\value{
The PSHD loss (nonnegative scalar) between the two supplied networks.
}
\description{
Penalized Structural Hamming Distance
}
