% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pshd.R
\name{estimate_pshd}
\alias{estimate_pshd}
\title{Find a netwwork structure that minimizes the expected PSHD loss among a set of samples}
\usage{
estimate_pshd(samples, a = 1)
}
\arguments{
\item{samples}{A list of length \eqn{B} or 3-dimensional array of dimensions \eqn{p \times p \times B} containing a set of \eqn{B} samples. Each sample should be a \eqn{p \times p} binary adjacency matrix representing the structure of a network.}

\item{a}{The value of the penalty parameter (scalar). Must be strictly greater than 0 and less than 2. Higher values (closer to 2) tend to return a more sparse final estimate with fewer edges, and vice versa for values closer to 0.}
}
\value{
A list containing two elements: The sample with minimal average PSHD loss to the rest of the samples, which is a binary \eqn{p \times p} adjacency matrix, and the value of that minimal expected loss.
}
\description{
Find a netwwork structure that minimizes the expected PSHD loss among a set of samples
}
