% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_pshd.R
\name{expected_pshd}
\alias{expected_pshd}
\title{Estimate the expected PSHD loss for a network given a set of samples}
\usage{
expected_pshd(g, samples, a = 1, indivLoss = FALSE)
}
\arguments{
\item{indivLoss}{}
}
\value{
xxxxx
}
\description{
Estimate the expected PSHD loss for a network given a set of samples
}
